added keep_temp_Rmd= argument

+++ R/make_sheet.R
@@ -11,11 +11,12 @@
 #' @param notes \code{logical}. If \code{handout=TRUE}, then if \code{TRUE} a writable box for notes will appear at the bottom of each slide.
 #' @param ntb \code{logical}. \code{TRUE} to render document as R Notebook. \code{FALSE} by default.
 #' @param color \code{character}. Either a single valid colour (hex code or any of the values in \code{colours()}) or any valid value of the \code{course=} argument. If provided, it will be used to set colour scheme instead of \code{course=}.
-#' @param toc \code{logical}. Should table of content be included
+#' @param toc \code{logical}. Should table of content be included.
 #' @param toc_depth \code{logical}. Depth of headers to include in table of contents.
 #' @param toc_float \code{TRUE} to float the table of contents to the left of the main document content. Rather than TRUE you may also pass a list of options that control the behavior of the floating table of contents. For more details, see \code{\link[rmarkdown]{html_document}}.
 #' @param fig_width,fig_depth \code{numeric}. Default width and height (in inches) for figures.
 #' @param highlight \code{character}. Syntax highlighting style. See \code{\link[rmarkdown]{html_document}}.
+#' @param keep_temp_Rmd \code{logical}. Should temporary Rmd file be kept? Useful for post-hoc edits and debugging. 
 #' @param ... Other arguments to be passed to \code{rmarkdown:html_document} or \code{rmarkdown:html_notebook}.
 #' @details Function requires a .css and .js files for correct formatting of lab sheets/handouts. These files sit on the stats website in the [root]/sheet_files folder and the path is hard-coded into the function. Look for css and js objects in function body.
 #' @return \code{TRUE} if output .html file was successfully created.
@@ -26,7 +27,7 @@
 
 make.sheet <- function(file, course, solution = F, handout = FALSE, tasks_to_headings = FALSE, notes = TRUE, ntb = FALSE,
                        color = NULL, toc = T, toc_depth = 2, toc_float = T, fig_width = 5, fig_height = 3.5,
-                       highlight = "tango", colour = color, ...) {
+                       highlight = "tango", colour = color, keep_temp_Rmd = F, ...) {
   if (!file.exists(file)) stop("The file does not exist.")
   if (!grepl("\\.[rR]md$", file)) stop("file= needs to be an .Rmd file.")
   if (!tolower(course) %in% c("dapr_1", "dapr_2", "dapr_3", "usmr", "msmr", "and", "ad", "adata", "other"))
@@ -166,7 +167,7 @@ make.sheet <- function(file, course, solution = F, handout = FALSE, tasks_to_hea
   
   temp_rmd <- sub("\\.[Rr]md$", "_temp.Rmd", file)
   writeLines(x, temp_rmd)
-  on.exit(file.remove(temp_rmd), add = T, after = F)
+  if (!keep_temp_Rmd) on.exit(file.remove(temp_rmd), add = T, after = F)
 
   out_html <- sub("\\.[Rr]md$", ifelse(solution, "_sol.html", ".html"), file)
